Refactor: extract UI helpers into separate files
Move and clean up several UI helper/data types into their own files for clarity and testability. Notable changes: ScheduleActions, RecentColors, RecentMediaFiles, RecentPictureFolders, FocusLostRescueState, SegmentedButtonItem, WindowInfo, ResolvedDisplay, SmartReference, DownloadState, RemoteActivityNotification, SendStatus and ShortcutCaptureTags were extracted into new files; duplicated / inlined definitions were removed from the original hosts (MainDesktop, ColorPickerDialog, FocusLostRescue, PicturesTab, MediaTab, RemoteActivityToast, UpdateAvailableDialog, ShortcutCaptureDialog). Several files were renamed for clarity (Platform.kt → JVMPlatform.kt, CommandOutput.kt → CommandResult.kt, DeckLinkIO.kt → DeckLinkManager.kt, ScheduleZoom.kt → ScheduleDensity.kt). Also enabled the detekt MatchingDeclarationName rule. These changes improve modularity and make state objects easier to unit-test.
