+++ composeApp/src/jvmMain/kotlin/org/churchpresenter/app/churchpresenter/composables/SegmentedButtonItem.kt
@@ -1,0 +1,10 @@
+package org.churchpresenter.app.churchpresenter.composables
+
+import androidx.compose.ui.graphics.vector.ImageVector
+
+data class SegmentedButtonItem<T>(
+    val value: T,
+    val label: String,
+    val tooltip: String? = null,
+    val icon: ImageVector? = null
+)
