+++ composeApp/src/jvmMain/kotlin/org/churchpresenter/app/churchpresenter/composables/WindowInfo.kt
@@ -1,0 +1,3 @@
+package org.churchpresenter.app.churchpresenter.composables
+
+internal data class WindowInfo(val title: String, val id: Long)
