+++ composeApp/src/jvmMain/kotlin/org/churchpresenter/app/churchpresenter/dialogs/ShortcutCaptureTags.kt
@@ -1,0 +1,9 @@
+package org.churchpresenter.app.churchpresenter.dialogs
+
+/** Test tags for the capture dialog's controls, which have no stable text to locate them by. */
+internal object ShortcutCaptureTags {
+    const val SURFACE = "shortcut_capture_surface"
+    const val PREVIEW = "shortcut_capture_preview"
+    const val CONFLICT = "shortcut_capture_conflict"
+    const val CONFIRM = "shortcut_capture_confirm"
+}
