+++ composeApp/src/jvmMain/kotlin/org/churchpresenter/app/churchpresenter/viewmodel/BibleViewModelSmartSearch.kt
@@ -11,13 +11,6 @@ import org.churchpresenter.app.churchpresenter.data.BibleSearch
  * acting on it.
  */
 
-internal data class SmartReference(
-    val bookIndex: Int,
-    val chapter: Int?,
-    val verseStart: Int?,
-    val verseEnd: Int?
-)
-
 internal fun BibleViewModel.canonicalBookIdToIndex(canonicalId: Int): Int? =
     _primaryBible.value?.getDisplayIndexForBookId(canonicalId)?.takeIf { it in _books.value.indices }
