+++ composeApp/src/jvmMain/kotlin/org/churchpresenter/app/churchpresenter/viewmodel/SmartReference.kt
@@ -1,0 +1,8 @@
+package org.churchpresenter.app.churchpresenter.viewmodel
+
+internal data class SmartReference(
+    val bookIndex: Int,
+    val chapter: Int?,
+    val verseStart: Int?,
+    val verseEnd: Int?
+)
